pyproject.toml
[tool.poetry]
name = "neomarkup"
version = "0.0.0"
description = "Never write a single line of HTML anymore! A modern markup language for the 21st century"
authors = ["baris-inandi <68742481+baris-inandi@users.noreply.github.com>"]
license = "GPLv2"

[tool.poetry.dependencies]
python = "^3.9"
rcssmin = "^1.1.0"
rjsmin = "^1.2.0"
libsass = "^0.21.0"
lesscpy = "^0.15.0"
six = "^1.16.0"
termcolor = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
